Added mount point for docker. Removed copy.
Ok so after some time I realized that error logs were not saving, and
that was obviously because the bot was writing in the docker container.
To fix this I removed the COPY thing from the Dockerfile and added a
mount point when running the bot with docker.sh

Probably not the best way but idk.